Added check to cleanDB to ensure model was not undefined

+++ server/src/seeds/cleanDb.ts
@@ -1,11 +1,17 @@
-import models from '../models/index.js';
-import db from '../config/connection.js';
+import models from "../models/index.js";
+import db from "../config/connection.js";
 
 export default async (modelName: "Question", collectionName: string) => {
   try {
-    let modelExists = await models[modelName].db.db.listCollections({
-      name: collectionName
-    }).toArray()
+    let modelTest = models[modelName].db.db;
+    if (!modelTest) {
+      throw new Error("Model not found");
+    }
+    let modelExists = await modelTest
+      .listCollections({
+        name: collectionName,
+      })
+      .toArray();
 
     if (modelExists.length) {
       await db.dropCollection(collectionName);
@@ -13,4 +19,4 @@ export default async (modelName: "Question", collectionName: string) => {
   } catch (err) {
     throw err;
   }
-}
+};
